Add ability to specify only json input file
Fides now has the ability for a json file to point itself to the .bp
file(s) containing the data. In this case, it makes sense to be able to
specify a json file with a bp file. This change allows you to specify the
`--json` input argument without `--file` to allow Fides to find its own bp
file.