Read file specified with `--file` as bp file
This was set to load the file as JSON, which would not work if it was a
bp file with an attribute containing the JSON metadata.

I expect this was set as a temporary thing to read in a specific json
file. However, there is already another change to allow you to specify
only a json file with `--json` and without `--file`.